Refactored drum pad settings
